requirements.txt
discord.py
riotwatcher
python-dotenv
requests
